+++ src/main/java/roomescape/exception/resource/AlreadyExistException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.resource;
 
 public class AlreadyExistException extends RuntimeException {
