+++ src/main/java/roomescape/member/application/MemberService.java
@@ -4,8 +4,8 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.auth.domain.AuthRole;
-import roomescape.exception.AuthorizationException;
-import roomescape.exception.ResourceNotFoundException;
+import roomescape.exception.auth.AuthorizationException;
+import roomescape.exception.resource.ResourceNotFoundException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberCommandRepository;
 import roomescape.member.domain.MemberQueryRepository;
