+++ src/main/java/roomescape/reservation/application/ReservationTimeService.java
@@ -6,8 +6,8 @@ import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
-import roomescape.exception.AlreadyExistException;
-import roomescape.exception.ResourceNotFoundException;
+import roomescape.exception.resource.AlreadyExistException;
+import roomescape.exception.resource.ResourceNotFoundException;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.ReservationTimeRepository;
 import roomescape.reservation.ui.dto.CreateReservationTimeRequest;
