+++ src/main/java/roomescape/theme/domain/ThemeRepository.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface ThemeRepository {
+    
     Long save(Theme theme);
 
     void deleteById(Long id);
